Authentication Changes
Logging in now generates a (custom) CSRF token, circumventing Laravels automatic Auth as it is designed to be added when the system is first created. The token is then used to figure out which pages a user should have access to. No bugs seem to be present with this yet. Also a couple of back button modifications.