Commenting + Bug Fixes + Specialist Page Updates
Began commenting controller files, started with ones which are unlikely to change. Fixed a few bugs whilst commenting that were remnants of old code. Started making it so specialist can actually view and edit problems, current approach is to use hasAccess() to render different views of the same route. Working well so far.