Finish Graph Basics
Graphing should now be much easier to roll out to other review pages, as it follows a 'object' oriented approach but using arrays and loops to use as many datasets as needed, without manually adding them to the JS. Equipment page and specialist page are 'done' in terms of having a basic graph. Software + Caller are main two left.