Start Specialist Holidays
Specialists can have holidays in the system and cannot receive problems when away, also added a flag to selector to warn operator if specialist is going away soon. Need to add same functionality to the problem creator, and have flag check it is within 7 or so days as a preliminary warning, and also make it so a specialist can book their holidays.